Reorganize commands and get rid of the ci command
