+++ xtask/src/commands/mod.rs
@@ -1,4 +1,5 @@
 pub(crate) mod books;
-pub(crate) mod checks;
-pub(crate) mod ci;
+pub(crate) mod build;
 pub(crate) mod doc;
+pub(crate) mod test;
+pub(crate) mod validate;
